fix trim() field

+++ controllers/question.controller.js
@@ -58,31 +58,41 @@ module.exports = {
     saveQuestion: (req, res) => {
         let qs_id = req.params.qs_id;
         let question = req.body;
-        question.content = question_model.format(question.content);
-        question.answer1 = question_model.format(question.answer1);
-        question.answer2 = question_model.format(question.answer2);
-        question.answer3 = question_model.format(question.answer3);
-        question.answer4 = question_model.format(question.answer4);
-        upload(req, res, err => {
-            if (err) {
-                req.flash("error", "Fail to create question!");
-                res.redirect(`/host/questionset/"${qs_id}"/question/add`);
-            } else {
-                var filename = "";
-                if (req.file) {
-                    filename = req.file.filename;
+        question.content = question_model.format(question.content).trim();
+        question.answer1 = question_model.format(question.answer1).trim();
+        question.answer2 = question_model.format(question.answer2).trim();
+        question.answer3 = question_model.format(question.answer3).trim();
+        question.answer4 = question_model.format(question.answer4).trim();
+        if (question.content.length <= 0 || question.answer1.length <= 0 || question.answer2.length <= 0 || question.answer3.length <= 0 || question.answer4.length <= 0) {
+            req.flash("error", "Begin is not a space");
+            res.render('questions/create-question', {
+                qs_id: qs_id,
+                user: req.user,
+                error: req.flash("error"),
+                csrfToken: req.csrfToken()
+            });
+        } else {
+            upload(req, res, err => {
+                if (err) {
+                    req.flash("error", "Fail to create question!");
+                    res.redirect(`/host/questionset/"${qs_id}"/question/add`);
+                } else {
+                    var filename = "";
+                    if (req.file) {
+                        filename = req.file.filename;
+                    }
+                    question_model.add(question, filename, qs_id)
+                        .then(result => {
+                            req.flash("success", "Create question successfully!");
+                            res.redirect(`/host/questionset/${qs_id}/question?page=1`);
+                        }).catch(err => {
+                            console.log(err);
+                            req.flash("error", "Fail to insert question!");
+                            res.redirect(`/host/questionset/${qs_id}/question/add`);
+                        });
                 }
-                question_model.add(question, filename, qs_id)
-                    .then(result => {
-                        req.flash("success", "Create question successfully!");
-                        res.redirect(`/host/questionset/${qs_id}/question?page=1`);
-                    }).catch(err => {
-                        console.log(err);
-                        req.flash("error", "Fail to insert question!");
-                        res.redirect(`/host/questionset/${qs_id}/question/add`);
-                    });
-            }
-        })
+            })
+        }
     },
     deleteQuestion: (req, res) => {
         // delete file
@@ -124,51 +134,58 @@ module.exports = {
                 }
                 res.render('questions/update', {
                     csrfToken: req.csrfToken(),
+                    error: req.flash("error"),
                     question: result[0],
                     user: req.user
                 })
             })
     },
     editQuestion: (req, res) => {
+        let qs_id = req.params.qs_id;
         let question = req.body;
-        question.content = question_model.format(question.content);
-        question.answer1 = question_model.format(question.answer1);
-        question.answer2 = question_model.format(question.answer2);
-        question.answer3 = question_model.format(question.answer3);
-        question.answer4 = question_model.format(question.answer4);
-        upload(req, res, err => {
-            if (err) {
-                console.log(err);
-                res.render('error')
-            } else {
-                let filename = "";
-                let sql = '';
-                if (req.file) {
-                    filename = req.file.filename;
-                    try {
-                        //file removed
-                        fs.unlink('./public/img/' + question.image, (err) => {
-                            if (err) {
-                                console.error(err)
-                                return
-                            }
-                        })
-                    } catch (err) {
-                        console.error(err)
-                    };
-
+        question.content = question_model.format(question.content).trim();
+        question.answer1 = question_model.format(question.answer1).trim();
+        question.answer2 = question_model.format(question.answer2).trim();
+        question.answer3 = question_model.format(question.answer3).trim();
+        question.answer4 = question_model.format(question.answer4).trim();
+        if (question.content.length <= 0 || question.answer1.length <= 0 || question.answer2.length <= 0 || question.answer3.length <= 0 || question.answer4.length <= 0) {
+            req.flash("error", "Begin is not a space");
+            res.redirect('back');
+        } else {
+            upload(req, res, err => {
+                if (err) {
+                    console.log(err);
+                    res.render('error')
                 } else {
-                    filename = question.image;
+                    let filename = "";
+                    let sql = '';
+                    if (req.file) {
+                        filename = req.file.filename;
+                        try {
+                            //file removed
+                            fs.unlink('./public/img/' + question.image, (err) => {
+                                if (err) {
+                                    console.error(err)
+                                    return
+                                }
+                            })
+                        } catch (err) {
+                            console.error(err)
+                        };
+
+                    } else {
+                        filename = question.image;
+                    }
+                    question_model.update(question, filename, req.params.q_id)
+                        .then(result => {
+                            res.redirect(`/host/questionset/${req.params.qs_id}/question?page=1`);
+                        })
+                        .catch(err => {
+                            console.log(err);
+                        });
                 }
-                question_model.update(question, filename, req.params.q_id)
-                    .then(result => {
-                        res.redirect(`/host/questionset/${req.params.qs_id}/question?page=1`);
-                    })
-                    .catch(err => {
-                        console.log(err);
-                    });
-            }
-        })
+            })
+        }
     },
 
 };
